website/python/requirements.txt
flask
flask-cors
torch
clip @ git+https://github.com/openai/CLIP.git
faiss-cpu
numpy
Pillow
opencv-python
insightface
tqdm
scikit-learn
